Chore: Delete contentNavStack
